refactor(spark): consolidate the nested schema pruning rule into hudi-spark-common
The six SparkXXNestedSchemaPruning subclasses of
BaseHoodieNestedSchemaPruning only bridged two Spark API deltas:
- StructType#toAttributes vs DataTypeUtils#toAttributes (SPARK-44353):
  replaced with inline AttributeReference construction, which is what
  both helpers do in every version
- the LogicalRelation unapply arity (a fifth field was added in Spark
  4.0): replaced with a type-based match plus a guard, which compiles
  against every version

With both deltas gone, the base class becomes the concrete
HoodieNestedSchemaPruning rule and HoodieAnalysis instantiates it
directly instead of reflectively picking one of six class-name strings.
